Splitting service operations based on service type #3253
